pyproject.toml
[tool.poetry]
name = "campus-cli"
version = "0.1.0"
description = "Campus CLI - Command-line interface for Campus API"
authors = ["NYJC Computing <nyjc.computing@nyjc.edu.sg>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
typer = {extras = ["all"], version = "^0.15.0"}
rich = "^13.9.0"
keyring = "^25.6.0"
campus-api-python = {git = "https://github.com/nyjc-computing/campus-api-python.git", branch = "main"}

[tool.poetry.scripts]
campus = "campus_cli.cli:app"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
ruff = "^0.9.0"

[[tool.poetry.packages]]
include = "campus_cli"

[tool.pyright]
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
